src/test/java/org/broadinstitute/hellbender/tools/funcotator/dataSources/XSV/XSVFuncotationUnitTest.java
package org.broadinstitute.hellbender.tools.funcotator.dataSources.XSV;

import org.broadinstitute.hellbender.GATKBaseTest;
import org.testng.Assert;
import org.testng.annotations.DataProvider;
import org.testng.annotations.Test;

import java.util.Arrays;
import java.util.Collections;

/**
 * A Unit Test class for {@link XSVFuncotation}
 * Created by jonn on 11/28/17.
 */
public class XSVFuncotationUnitTest extends GATKBaseTest {

    //==================================================================================================================
    // Private Static Members:

    //==================================================================================================================
    // Private Members:

    //==================================================================================================================
    // Helper Methods:

    //==================================================================================================================
    // Data Providers:

    @DataProvider
    Object[][] provideForTestGet() {
        return new Object[][]{
                { "X" },
                { "A" },
                { "B" },
                { "Y" },
                { "L" },
                { "R" },
                { "Select" },
                { "Start" },
        };
    }

    @DataProvider
    Object[][] provideDataForTestSetFieldSerializationOverrideValue() {

        final XSVFuncotation funcotation =
                new XSVFuncotation(
                    Arrays.asList("A", "B", "C"), Arrays.asList("1", "2", "3")
                );

        return new Object[][] {
                { funcotation, "A", null },
                { funcotation, "A", "A" },
                { funcotation, "A", "B" },
                { funcotation, "A", "C" },
                { funcotation, "B", "A" },
                { funcotation, "B", "B" },
                { funcotation, "B", "C" },
                { funcotation, "C", "A" },
                { funcotation, "C", "B" },
                { funcotation, "C", "C" },
        };
    }

    @DataProvider
    Object[][] provideForTestSerializeToVcfString() {
        return new Object[][] {
                {
                    new XSVFuncotation(Collections.emptyList(), Collections.emptyList()),
                    ""
                },
                {
                    new XSVFuncotation(Collections.singletonList("A"), Collections.singletonList(("1"))),
                    "1"
                },
                {
                    new XSVFuncotation(Arrays.asList("A", "B"), Arrays.asList("1", "2")),
                    "1|2"
                },
                {
                    new XSVFuncotation(Arrays.asList("A", "B", "C"), Arrays.asList("1", "2", "3")),
                    "1|2|3"
                },
        };
    }

    @DataProvider
    Object[][] provideListOfStrings() {
        return new Object[][] {
                { Collections.singletonList("A") },
                { Arrays.asList("A", "B") },
                { Arrays.asList("A", "B", "C") },
                { Arrays.asList("A", "B", "C", "D") },
                { Arrays.asList("A", "B", "C", "D", "E") },
        };
    }

    //==================================================================================================================
    // Tests:

    @Test(dataProvider = "provideForTestGet")
    public void testGet(final String fieldValue) {

        final String fieldName = "PLACEHOLDER";

        final XSVFuncotation funcotation = new XSVFuncotation( Collections.singletonList(fieldName), Collections.singletonList(fieldValue) );
        Assert.assertEquals( funcotation.get(fieldName), fieldValue );
    }

    @Test(dataProvider = "provideDataForTestSetFieldSerializationOverrideValue")
    public void testSetFieldSerializationOverrideValue(final XSVFuncotation funcotation,
                                                       final String fieldName,
                                                       String overrideValue) {
        if ( overrideValue == null ) {
            overrideValue = funcotation.get(fieldName);
        }
        else {
            funcotation.setFieldSerializationOverrideValue(fieldName, overrideValue);
        }
        Assert.assertEquals( funcotation.get(fieldName), overrideValue );
    }

    @Test(dataProvider = "provideForTestSerializeToVcfString")
    public void testSerializeToVcfString(final XSVFuncotation funcotation,
                                         final String expected) {
        Assert.assertEquals( funcotation.serializeToVcfString(), expected );
    }

//    @Test(dataProvider = "provideListOfStrings")
//    public void test

}
